% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErlangFuncs.R
\name{Erlang_NegLogLikelihood}
\alias{Erlang_NegLogLikelihood}
\title{Negative Log-Likelihood for Erlang Distribution}
\usage{
Erlang_NegLogLikelihood(k, data)
}
\arguments{
\item{k}{Numeric, Erlang shape parameter.}

\item{data}{Numeric vector of observed data.}
}
\value{
Numeric scalar representing the negative log-likelihood.
}
\description{
Computes the negative log-likelihood (to be minimized) for a given
Erlang (Gamma) shape parameter \code{k} and data vector. The scale parameter
is internally estimated as \eqn{lambda = mean(data) / k}.
}
\examples{
\dontrun{
data <- rexp(200, rate = 0.5)
Erlang_NegLogLikelihood(3, data)
}

}
\seealso{
\code{\link[=Erlang_Fit_v2]{Erlang_Fit_v2()}}, \code{\link[=Erlang_Fit_v2_Pvalue]{Erlang_Fit_v2_Pvalue()}}
}
\keyword{internal}
