% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenErlangFit-package.R
\docType{package}
\name{GenErlangFit-package}
\alias{GenErlangFit}
\alias{GenErlangFit-package}
\title{GenErlangFit: Fitting Erlang and Erlang–Exponential Time-to-Event Distributions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions for fitting Erlang and Erlang-Exponential distributions using maximum likelihood estimation, including support for multi-K window search, plotting, and diagnostics.
}
\author{
\strong{Maintainer}: Vivek Murali \email{vivekmurals@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Alison Hill [copyright holder]
  \item Anjalika Nande [copyright holder]
}

}
\keyword{internal}
